feat: use more precise type for uri & datetime than String
- uri => fluent_uri::Uri<String>
- uri-referece => fluent_uri::Uri<String>
- datetime => time::OffsetDateTime
