+++ cdevents-sdk/Cargo.toml
@@ -11,9 +11,11 @@ publish = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-http = "1"
-http-serde = "2"
+# http = "1"                                                      # doesn't support uri-reference
+# http-serde = "2"
+fluent-uri = "0.1"
 serde = { version = "1.0", features = ["derive"] }
+serde_with = "3.4"
 serde_json = "1.0"
 time = { version = "0.3", features = ["serde-human-readable"] }
